Move Security classes and added field priority to DinaRole
