+++ dina-base-api/src/main/java/ca/gc/aafc/dina/security/DinaAuthorizationService.java
@@ -1,4 +1,4 @@
-package ca.gc.aafc.dina.service;
+package ca.gc.aafc.dina.security;
 
 public interface DinaAuthorizationService {
