+++ dina-base-api/src/main/java/ca/gc/aafc/dina/security/PriorityComparator.java
@@ -1,0 +1,12 @@
+package ca.gc.aafc.dina.security;
+
+import java.util.Comparator;
+
+public class PriorityComparator implements Comparator<DinaRole> {
+
+  @Override
+  public int compare(DinaRole firstDinaRole, DinaRole secondDinaRole) {
+    return Integer.compare(firstDinaRole.getPriority(), secondDinaRole.getPriority());
+  }
+  
+}
